# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "spelunk-server",
    "description": "Shared memory server for spelunk. Stores decisions, requirements, and context for a team and serves them over HTTP.",
    "contact": {
      "name": "spelunk",
      "url": "https://github.com/spelunk-cloud/spelunk"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}/memory": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "List memory entries for a project, optionally filtered by kind.",
        "operationId": "list_notes",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "Filter by kind (`decision`, `requirement`, `note`, `question`, `handoff`, `intent`).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return (default: 20).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "archived",
            "in": "query",
            "description": "Include archived entries (default: false).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerNote"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Add a memory entry to a project. The project is auto-created on first write.",
        "description": "The `embedding` field is optional. If omitted and the server's embedder is ready,\nthe server embeds the entry before storage. If neither is available, the\nentry is stored without a vector (text search only, no KNN).\n\nReturns **201** on success. Returns **409** when the new entry is semantically\nclose to one or more existing active entries (similarity ≥ conflict_threshold).\nThe entry is still stored in both cases; the 409 is informational.\nReturns **422** when the entry contains prompt-injection patterns.",
        "operationId": "add_note",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug (e.g. `usercise/spelunk`)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Note created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Embedding dimension mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Note stored but conflicts with existing entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Entry rejected: prompt injection detected"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/projects/{project_id}/memory/search": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Semantic search over memory entries. The server encodes the text query using its\nconfigured embedder. Returns 400 if no embedder is configured.",
        "operationId": "search_notes",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Nearest neighbours",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerNote"
                  }
                }
              }
            }
          },
          "400": {
            "description": "No embedder configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Embed admission queue full; retry after the given delay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/projects/{project_id}/memory/{note_id}": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Get a single memory entry by ID.",
        "operationId": "get_note",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerNote"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "memory"
        ],
        "summary": "Delete a memory entry permanently.",
        "operationId": "delete_note",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoolResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/memory/{note_id}/archive": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Archive a memory entry. Archived entries are excluded from search and `ask`\ncontext but remain visible via `?archived=true`.",
        "operationId": "archive_note",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archive result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoolResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/memory/{note_id}/supersede": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Mark a memory entry as superseded by a newer one. The old entry is archived\nand linked to the new one.",
        "operationId": "supersede_note",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "description": "Note ID to supersede",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupersedeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Supersede result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoolResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Embed a search query server-side and return the vector for the CLI to use\nin its local KNN search.",
        "description": "The server applies the F2LLM code-retrieval query prefix\n(`Instruct: Given a code search query…\\nQuery: {q}`) so the CLI does not\nneed to know the embedding format. The server does **not** perform the KNN\nstep: the local SQLite index lives on the CLI side.\n\n- `\"semantic\"` / `\"hybrid\"`: embeds the query and returns `query_vector`.\n  Returns **400** if no embedder is configured on this server.\n- `\"text\"`: returns `query_vector: null`; the CLI falls back to FTS.",
        "operationId": "project_search",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Query vector (CLI runs KNN locally)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "No embedder configured or invalid mode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Embed admission queue full; retry after the given delay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddNoteRequest": {
        "type": "object",
        "required": [
          "kind",
          "title"
        ],
        "properties": {
          "body": {
            "type": "string"
          },
          "embedding": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "Pre-computed embedding vector from the client. Optional: if omitted and the\nserver's embedder is ready, the server embeds the entry.\nIf neither is available, the entry is stored without a vector (text search only)."
          },
          "kind": {
            "type": "string",
            "description": "Kind of memory entry: `decision`, `requirement`, `note`, `question`, `handoff`, `intent`."
          },
          "linked_files": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Source file paths this entry is linked to."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional tags for filtering."
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Consistent JSON error body: `{\"error\": {\"code\": \"...\", \"message\": \"...\"}}`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Pass as `Authorization: Bearer <key>`. Not required when no key is configured on the server."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}